factor out logs, datetime, and usb functions from settings page
