+++ apps/pollbook/frontend/src/nav_screen.tsx
@@ -475,6 +475,7 @@ export const electionManagerRoutes = {
   voters: { title: 'Voters', path: '/voters' },
   addVoter: { title: 'Registration', path: '/registration' },
   statistics: { title: 'Statistics', path: '/statistics' },
+  settings: { title: 'Settings', path: '/settings' },
 } satisfies Record<string, { title: string; path: string }>;
 
 export function ElectionManagerNavScreen({
