+++ apps/pollbook/frontend/src/settings_screen.tsx
@@ -1,0 +1,53 @@
+import {
+  CurrentDateAndTime,
+  ExportLogsButton,
+  FormatUsbButton,
+  H2,
+  MainContent,
+  P,
+  SetClockButton,
+} from '@votingworks/ui';
+import { formatUsbDrive, getUsbDriveStatus, logOut } from './api';
+
+interface SettingsScreenProps {
+  showFormatUsbButton: boolean;
+}
+
+export function SettingsScreen({
+  showFormatUsbButton,
+}: SettingsScreenProps): JSX.Element | null {
+  const logOutMutation = logOut.useMutation();
+  const usbDriveStatusQuery = getUsbDriveStatus.useQuery();
+  const formatUsbDriveMutation = formatUsbDrive.useMutation();
+
+  if (!usbDriveStatusQuery.isSuccess) {
+    return null;
+  }
+
+  const usbDriveStatus = usbDriveStatusQuery.data;
+
+  return (
+    <MainContent>
+      <H2>Logs</H2>
+      <ExportLogsButton usbDriveStatus={usbDriveStatus} />
+      <H2>Date and Time</H2>
+      <P>
+        <CurrentDateAndTime />
+      </P>
+      <P>
+        <SetClockButton logOut={() => logOutMutation.mutate()}>
+          Set Date and Time
+        </SetClockButton>
+      </P>
+      <H2>USB</H2>
+      {showFormatUsbButton && (
+        <P>
+          <FormatUsbButton
+            usbDriveStatus={usbDriveStatus}
+            formatUsbDriveMutation={formatUsbDriveMutation}
+          />
+        </P>
+      )}
+    </MainContent>
+  );
+}
